+++ src/components/ui/Icon.svelte
@@ -34,6 +34,7 @@
   import Telescope from '@lucide/svelte/icons/telescope';
   import GraduationCap from '@lucide/svelte/icons/graduation-cap';
   import ChartColumn from '@lucide/svelte/icons/chart-no-axes-column';
+  import ChartGantt from '@lucide/svelte/icons/chart-no-axes-gantt';
   import Ruler from '@lucide/svelte/icons/ruler';
   import TreePine from '@lucide/svelte/icons/tree-pine';
   import Landmark from '@lucide/svelte/icons/landmark';
@@ -83,6 +84,7 @@
     telescope: Telescope,
     'graduation-cap': GraduationCap,
     chart: ChartColumn,
+    'chart-no-axes-gantt': ChartGantt,
     ruler: Ruler,
     'tree-pine': TreePine,
     landmark: Landmark,